+++ server/ctrlsubsonic/spec/construct_by_tags.go
@@ -1,6 +1,7 @@
 package spec
 
 import (
+	"cmp"
 	"path/filepath"
 	"sort"
 
@@ -22,6 +23,8 @@ func NewAlbumByTags(a *db.Album, artists []*db.Artist) *Album {
 		Year:          a.TagYear,
 		Tracks:        []*TrackChild{},
 		AverageRating: formatRating(a.AverageRating),
+		IsCompilation: a.TagCompilation,
+		ReleaseTypes:  formatReleaseTypes(a.TagReleaseType),
 	}
 	if a.Cover != "" {
 		ret.CoverID = a.SID()
@@ -76,7 +79,7 @@ func NewTrackByTags(t *db.Track, album *db.Album) *TrackChild {
 		Path:               filepath.Join(album.LeftPath, album.RightPath, t.Filename),
 		Size:               t.Size,
 		Suffix:             formatExt(t.Ext()),
-		Title:              t.TagTitle,
+		Title:              cmp.Or(t.TagTitle, t.Filename),
 		TrackNumber:        t.TagTrackNumber,
 		DiscNumber:         t.TagDiscNumber,
 		Type:               "music",